+++ src/layout/pages/seaPage/index.tsx
@@ -1,0 +1,1 @@
+export { SeaPage, SeaPageProps, IllustrationType } from './seaPage'