+++ src/marquee/index.tsx
@@ -1,0 +1,1 @@
+export { Marquee, MarqueeProps } from './Marquee'