+++ src/typography/index.tsx
@@ -4,6 +4,8 @@ import { color } from '../_utils/branding'
 import { Text } from './Text'
 
 const StyledText = styled(Text)`
+  white-space: pre-line;
+
   & {
     color: ${props => (props.isInverted ? color.white : '')};
   }